lazy load MaterialDesignComponents
